+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   `Popover`, `Dialog` & `Lightbox`: fix close on click away inside a shadow DOM
+
 ## [3.17.1][] - 2025-09-11
 
 ### Fixed